refactor: insertImage페이지 컴포넌트로 변경 후 적용
