+++ frontend/src/routes.tsx
@@ -1,5 +1,5 @@
 import React, { lazy } from 'react';
-import InsertImage from './page/insertImage';
+import InsertImage from './components/insertImage';
 import GlobalLayout from './page/_layout';
 
 const Main = lazy(() => import('./page/main'));
@@ -11,7 +11,6 @@ export const routes = [
     element: <GlobalLayout />,
     children: [
       { path: '/', element: <Main />, index: true },
-      { path: '/insert', element: <InsertImage></InsertImage>, index: true },
       { path: '/result', element: <Result />, index: true },
       { path: '/fishList', element: <FishList></FishList>, index: true },
     ],
